% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sarah-Seq.R
\name{vcf_to_23}
\alias{vcf_to_23}
\title{Convert vcf to 23andMe format}
\usage{
vcf_to_23(input_filepath, return_df = F, zip = T)
}
\arguments{
\item{input_filepath}{The path to the vcf file that you want to convert.}

\item{return_df}{If FALSE, the function will just convert the vcf and write out the resulting file in 23andMe format. If TRUE, no file is written but the 23andMe format dataframe is returned. Defaults to FALSE.}

\item{zip}{If TRUE, the 23andMe txt file that's created will be zipped to reduce file size. Defaults to TRUE.}
}
\description{
This function takes in a vcf file and converts it to 23andMe format so it can be uploaded to third party apps such as \href{https://www.impute.me}{impute.me}.
}
\examples{
vcf_to_23('my_vcf.vcf')
my23_df <- vcf_to_23('my_vcf.vcf', return_df = T)
}
